Code - Fix

+++ Classes/IpCamera.php
@@ -83,6 +83,10 @@ class IpCamera {
                         $this->fileAction($parameters);
                     else if ($_GET['controller'] == "userProfilePasswordAction")
                         $this->userProfilePasswordAction($parameters);
+                    else if ($_GET['controller'] == "userManagementSelectionAction")
+                        $this->userManagementSelectionAction($parameters);
+                    else if ($_GET['controller'] == "userManagementProfileAction")
+                        $this->userManagementProfileAction($parameters);
                     else if ($_GET['controller'] == "settingAction")
                         $this->settingAction($parameters);
                 }
@@ -119,6 +123,15 @@ class IpCamera {
                     }
                 }
             }
+        }
+    }
+    
+    public function generateSelectOptionUser() {
+        $userRows = $this->query->selectAllUserDatabase(1);
+        
+        if ($userRows != false) {
+            foreach ($userRows as $key => $value)
+                echo "<option value=\"{$value['id']}\">{$value['username']}</option>";
         }
     }
     
@@ -189,8 +202,6 @@ class IpCamera {
     
     // Functions private
     private function curlCommandsUrls($url) {
-        $response = "";
-        
         $curl = curl_init();
         
         curl_setopt($curl, CURLOPT_URL, $url);
@@ -210,7 +221,7 @@ class IpCamera {
         
         $deviceRow = $this->query->selectDeviceDatabase($deviceId);
         
-        $content = "framerate 30\n";
+        $content = "framerate 10\n";
         $content .= "netcam_url $videoUrl/{$deviceRow['video']}user=$username&pwd=$password&resolution=$this->resolution\n";
         $content .= "netcam_userpass $username:$password\n";
         $content .= "threshold $threshold\n";
@@ -231,7 +242,7 @@ class IpCamera {
         @chmod("{$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}.conf", 0664);
         
         $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?target_dir={$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}");
-        $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?framerate=30");
+        $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?framerate=10");
         $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?netcam_url=$videoUrl/{$deviceRow['video']}user=$username&pwd=$password&resolution=$this->resolution");
         $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?netcam_userpass=$username:$password");
         $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/config/set?threshold=$threshold");
@@ -251,41 +262,39 @@ class IpCamera {
     
     // Controllers
     private function selectionAction($parameters) {
+        $this->utility->checkSessionOverTime();
+        
         if (isset($parameters['cameraNumber']) == false)
             return;
         
-        if ($parameters['cameraNumber'] == 0) {
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($parameters['cameraNumber'] == 0 && $checkRoleUser == true) {
             $cameraRows = $this->query->selectAllCameraDatabase();
             
             $lastCamera = end($cameraRows);
-            $_SESSION['camera_number'] = $lastCamera['camera_number'] + 1;
+            $_SESSION['camera_number'] = $lastCamera['number'] + 1;
             
             $query = $this->utility->getDatabase()->getPdo()->prepare("INSERT INTO cameras (
-                                                                            camera_number,
+                                                                            number,
                                                                             device_id,
                                                                             video_url,
                                                                             username,
-                                                                            password,
-                                                                            threshold,
-                                                                            motion_detection_status
+                                                                            password
                                                                         )
                                                                         VALUES (
-                                                                            :cameraNumber,
+                                                                            :number,
                                                                             :deviceId,
                                                                             :videoUrl,
                                                                             :username,
-                                                                            :password,
-                                                                            :threshold,
-                                                                            :motionDetectionStatus
+                                                                            :password
                                                                         );");
             
-            $query->bindValue(":cameraNumber", $_SESSION['camera_number']);
+            $query->bindValue(":number", $_SESSION['camera_number']);
             $query->bindValue(":deviceId", "");
             $query->bindValue(":videoUrl", "");
             $query->bindValue(":username", "");
             $query->bindValue(":password", "");
-            $query->bindValue(":threshold", "");
-            $query->bindValue(":motionDetectionStatus", "pause");
             
             $query->execute();
             
@@ -313,6 +322,8 @@ class IpCamera {
     
     private function controlAction($parameters) {
         if ($parameters['event'] == "picture") {
+            $this->utility->checkSessionOverTime();
+            
             $this->curlCommandsUrls("{$this->settingRow['server_url']}/{$_SESSION['camera_number']}/action/snapshot");
             
             $this->response['messages']['success'] = "Picture taked.";
@@ -320,24 +331,33 @@ class IpCamera {
     }
     
     private function apparatusProfileAction($parameters) {
-        $this->apparatusProfileConfig($parameters['deviceId'], $parameters['videoUrl'], $parameters['username'], $parameters['password'], $parameters['threshold'], $parameters['motionDetectionStatus']);
+        $this->utility->checkSessionOverTime();
+        
+        $this->apparatusProfileConfig($parameters['deviceId'],
+                                        $parameters['videoUrl'],
+                                        $parameters['username'],
+                                        $parameters['password'],
+                                        $parameters['threshold'],
+                                        $parameters['motionDetectionStatus']);
         
         $query = $this->utility->getDatabase()->getPdo()->prepare("UPDATE cameras
-                                                                    SET device_id = :deviceId,
+                                                                    SET label = :label,
+                                                                        device_id = :deviceId,
                                                                         video_url = :videoUrl,
                                                                         username = :username,
                                                                         password = :password,
                                                                         threshold = :threshold,
                                                                         motion_detection_status = :motionDetectionStatus
-                                                                    WHERE camera_number = :cameraNumber");
+                                                                    WHERE number = :number");
         
+        $query->bindValue(":label", $parameters['label']);
         $query->bindValue(":deviceId", $parameters['deviceId']);
         $query->bindValue(":videoUrl", $parameters['videoUrl']);
         $query->bindValue(":username", $parameters['username']);
         $query->bindValue(":password", $parameters['password']);
         $query->bindValue(":threshold", $parameters['threshold']);
         $query->bindValue(":motionDetectionStatus", $parameters['motionDetectionStatus']);
-        $query->bindValue(":cameraNumber", $_SESSION['camera_number']);
+        $query->bindValue(":number", $_SESSION['camera_number']);
         
         $query->execute();
         
@@ -347,6 +367,13 @@ class IpCamera {
     }
     
     private function apparatusProfileDeleteAction() {
+        $this->utility->checkSessionOverTime();
+        
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($checkRoleUser == false)
+            return;
+        
         $this->utility->removeDirRecursive("{$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}", true);
         
         unlink("{$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}.conf");
@@ -354,9 +381,9 @@ class IpCamera {
         $this->utility->searchInFile("/etc/motion/motion.conf", "thread {$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}.conf", " ");
         
         $query = $this->utility->getDatabase()->getPdo()->prepare("DELETE FROM cameras
-                                                                    WHERE camera_number = :cameraNumber");
+                                                                    WHERE number = :number");
         
-        $query->bindValue(":cameraNumber", $_SESSION['camera_number']);
+        $query->bindValue(":number", $_SESSION['camera_number']);
         
         $query->execute();
         
@@ -368,6 +395,13 @@ class IpCamera {
     }
     
     private function fileAction($parameters) {
+        $this->utility->checkSessionOverTime();
+        
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($checkRoleUser == false)
+            return;
+        
         if ($parameters['event'] == "delete") {
             $path = "{$_SERVER['DOCUMENT_ROOT']}/motion/camera_{$_SESSION['camera_number']}/" . trim($parameters['name']);
             
@@ -388,7 +422,9 @@ class IpCamera {
     }
     
     private function userProfilePasswordAction($parameters) {
-        $messagePassword = $this->ipCameraUtility->assigUserPassword("withOld", $_SESSION['userLogged'], $parameters);
+        $this->utility->checkSessionOverTime();
+        
+        $messagePassword = $this->ipCameraUtility->assigUserPassword("withOld", $_SESSION['user_logged'], $parameters);
 
         if (isset($messagePassword['password']) == true) {
             $query = $this->utility->getDatabase()->getPdo()->prepare("UPDATE users
@@ -396,7 +432,7 @@ class IpCamera {
                                                                         WHERE id = :id");
 
             $query->bindValue(":password", $messagePassword['password']);
-            $query->bindValue(":id", $_SESSION['userLogged']['id']);
+            $query->bindValue(":id", $_SESSION['user_logged']['id']);
 
             $query->execute();
 
@@ -406,7 +442,111 @@ class IpCamera {
             $this->response['messages']['error'] = $messagePassword['error'];
     }
     
+    private function userManagementSelectionAction($parameters) {
+        $this->utility->checkSessionOverTime();
+        
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($checkRoleUser == false)
+            return;
+        
+        $_SESSION['user_management'] = null;
+        
+        if ($parameters['id'] == 0) {
+            $_SESSION['user_management'] = "new";
+            
+            $this->response['render'] = "userManagement_profile.php";
+        }
+        else if ($parameters['id'] > 0) {
+            $userRow = $this->query->selectUserDatabase($parameters['id']);
+            
+            $_SESSION['user_management'] = $userRow;
+            
+            $this->response['render'] = "userManagement_profile.php";
+        }
+        else
+            $this->response['messages']['error'] = "No user selected!";
+    }
+    
+    private function userManagementProfileAction($parameters) {
+        $this->utility->checkSessionOverTime();
+        
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($checkRoleUser == false)
+            return;
+        
+        if (isset($_SESSION['user_management']) == true) {
+            $messagePassword = $this->ipCameraUtility->assigUserPassword("withoutOld", $_SESSION['user_management'], $parameters);
+            
+            if (isset($messagePassword['password']) == true) {
+                if ($_SESSION['user_management'] == "new") {
+                    $query = $this->utility->getDatabase()->getPdo()->prepare("INSERT INTO users (
+                                                                                    role_user_id,
+                                                                                    username,
+                                                                                    email,
+                                                                                    password,
+                                                                                    not_locked
+                                                                                )
+                                                                                VALUES (
+                                                                                    :roleUserId,
+                                                                                    :username,
+                                                                                    :email,
+                                                                                    :password,
+                                                                                    :notLocked
+                                                                                );");
+
+                    $query->bindValue(":roleUserId", $parameters['roleUserId']);
+                    $query->bindValue(":username", $parameters['username']);
+                    $query->bindValue(":email", $parameters['email']);
+                    $query->bindValue(":password", $messagePassword['password']);
+                    $query->bindValue(":notLocked", "1");
+
+                    $query->execute();
+
+                    $this->response['messages']['success'] = "User created with success.";
+                }
+                else {
+                    $query = $this->utility->getDatabase()->getPdo()->prepare("UPDATE users
+                                                                                SET role_user_id = :roleUserId,
+                                                                                    username = :username,
+                                                                                    email = :email,
+                                                                                    password = :password,
+                                                                                    not_locked = :notLocked
+                                                                                WHERE id = :id");
+
+                    $query->bindValue(":roleUserId", $parameters['roleUserId']);
+                    $query->bindValue(":username", $parameters['username']);
+                    $query->bindValue(":email", $parameters['email']);
+                    
+                    if ($messagePassword['password'] == "")
+                        $query->bindValue(":password", $_SESSION['user_management']['password']);
+                    else
+                        $query->bindValue(":password", $messagePassword['password']);
+                    
+                    $query->bindValue(":notLocked", $parameters['notLocked']);
+                    $query->bindValue(":id", $_SESSION['user_management']['id']);
+
+                    $query->execute();
+                    
+                    $this->response['messages']['success'] = "User updated with success.";
+                }
+            }
+            else
+                $this->response['messages']['error'] = $messagePassword['error'];
+        }
+        else
+            $this->response['messages']['error'] = "User not create!";
+    }
+    
     private function settingAction($parameters) {
+        $this->utility->checkSessionOverTime();
+        
+        $checkRoleUser = $this->ipCameraUtility->checkRoleUser(Array("ROLE_ADMIN"), $_SESSION['user_logged']['role_user_id']);
+        
+        if ($checkRoleUser == false)
+            return;
+        
         $query = $this->utility->getDatabase()->getPdo()->prepare("UPDATE settings
                                                                     SET template = :template,
                                                                         server_url = :serverUrl,
